CBM: Final changes to VHD file
The code still doesn't fully work, but this commit adds the final changes
to the logic prior to handing over the project.

A document about the logic will be edited; this document will contain
details about the functioning of the logic, as well as details on "how" it
doesn't work.